requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
websockets==10.0
redis==4.3.4